[fixed] shouldFocusAfterRender and shouldReturnFocusAfterClose flags.
Both options should work independent.

If shouldFocusAfterRender and shouldReturnFocusAfterClose are true,
the default behavior, it should focus the modal when open and return
the focus to the previous element when close.

If shouldFocusAfterRender is false and shouldReturnFocusAfterClose is true
should be the same as both option true - noop.

If shouldFocusAfterRender is true and shouldReturnFocusAfterClose is false,
then we focus the modal when open and on close we just pop the previous active
element from the focus manager.